Fix infinite loop in `readExpression()`.
While comparing operators for precendence on the stack on what is
now calc.jsL315 we would fall into an infinite loop if the token
being compared had greater precendence. An example that caused this
behavior is:

1 + 2 - 3 / 4 * 5

When reading the expression we fell into an infinite loop upon reaching
the "/" character. At this point the "-" would be the only thing on the
stack and "/" has higher precedence than "-" so "-" would not be popped
off the stack, thus the stack would not change in size so the while
loop powering precendence would never exit.

This update simplifies the code by removing an if statement, instead
moving the code to be part of the while statement.